fix: widen before multiplying cell counts, not after
Eight places multiplied two `u16`s and widened the result, so each one
wrapped instead of producing the value it was cast to.

The reachable one is in the halfblocks encoder. `encode` allocates

    (size.width * size.height) as usize

and then writes to it at

    x + (size.width as usize) * (y / 2)

computed in `usize`. Below 65535 cells the two agree exactly. One cell
past it the buffer is allocated modulo 65536 while the loop still walks
the true extent, so the encoder indexes past its own end and panics with
a length that looks arbitrary. Reported from a 458x144 pane: 65952 cells,
allocated as 416, panic at index 416.

The threshold is a plausible terminal today. 458x144 is a 460x147 window
with a one-cell border, which is an ordinary full-screen terminal at a
small font -- and halfblocks is the backend most likely to be asked for a
very fine grid, since its resolution IS the cell grid.

The other seven are the same shape in `cells * font_size`:

  * `Resize::resize` and `needs_resize` (src/lib.rs) size an image in
    pixels from a cell count. These do not panic -- they silently resize
    to a wrapped pixel dimension once the product passes 65535, which a
    large image reaches easily.
  * `slice_rows` (src/sliced.rs) does the same before `.into()`.

Fixed uniformly by widening the operands rather than the product. No
behaviour changes below the ceiling and no API changes.

The regression test asserts one entry per cell at 458x144. It fails
before this change in both profiles, with different faces: in debug,
overflow-checks fire on the multiply itself; in release it wraps and the
panic is the out-of-bounds index above.

+++ src/lib.rs
@@ -413,8 +413,8 @@ impl Resize {
         size: Size,
         background_color: Option<Rgba<u8>>,
     ) -> DynamicImage {
-        let width = (size.width * font_size.width) as u32;
-        let height = (size.height * font_size.height) as u32;
+        let width = u32::from(size.width) * u32::from(font_size.width);
+        let height = u32::from(size.height) * u32::from(font_size.height);
 
         // Resize/Crop/etc., fitting a multiple of font-size, but not necessarily the `size`.
         let mut image = self.resize_pixels(image, width, height);
@@ -470,8 +470,8 @@ impl Resize {
             && desired.height <= target.height
             && (current.is_none() || current == Some(desired))
         {
-            let width = (desired.width * font_size.width) as u32;
-            let height = (desired.height * font_size.height) as u32;
+            let width = u32::from(desired.width) * u32::from(font_size.width);
+            let height = u32::from(desired.height) * u32::from(font_size.height);
             if image.width() == width || image.height() == height {
                 return None;
             }
